requirements.txt
Django
djangorestframework
drf-yasg
gunicorn
djangorestframework-simplejwt
psycopg2-binary
python-decouple
django-filter